requirements.txt
django
django-filter
djangorestframework
whitenoise
